Simple Being creation and trait matching.
So, if amount of traits tracked is less and the number of traits to pick from is also less, we can have higher compatibility.
But as soon as the number of traits increase per being and the gamut to choose from also increases, we get lesser compatibility.
These are cursory remarks.